Expose ports 2888 and 3888 so zookeeper instances could communicate to each other.
From the zookeeper doc:
Finally, note the two port numbers after each server name: " 2888" and "3888". Peers use the former port to connect to other peers. Such a connection is necessary so that peers can communicate, for example, to agree upon the order of updates. More specifically, a ZooKeeper server uses this port to connect followers to the leader. When a new leader arises, a follower opens a TCP connection to the leader using this port. Because the default leader election also uses TCP, we currently require another port for leader election. This is the second port in the server entry.

Signed-off-by: Diyan Yordanov <dgyordanov@gmail.com>